add listeners only can edit their own ratings and error messages

+++ client/src/RatingCard.js
@@ -8,25 +8,31 @@ function RatingCard({rating, song, setRatings, listener}) {
     for(let i=0; i< rating.review ; i++){
         stars += star
     }
-
-    //make it so listener can only change their own ratings
-
+    const [editError, setEditError] = useState(false)
     const [commentChange, setCommentChange] = useState("")
 
     function handleDelete(e) {
         e.preventDefault()
-        fetch(`/songs/${song.id}/ratings/${rating.id}`, {
-            method: 'DELETE',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-            body: JSON.stringify(null), 
-        })
-        .then(resp => console.log(resp))
-        .then(() => {
-            const newRatings = song.ratings.filter(item => item.id != rating.id)
-            setRatings(newRatings)
-        })
+        if (listener.id == rating.id) {
+            fetch(`/songs/${song.id}/ratings/${rating.id}`, {
+                method: 'DELETE',
+                headers: {
+                    'Content-Type': 'application/json',
+                },
+                body: JSON.stringify(null), 
+            })
+            .then(resp => console.log(resp))
+            .then(() => {
+                const newRatings = song.ratings.filter(item => item.id != rating.id)
+                setRatings(newRatings)
+            })
+        }
+        else {
+            setEditError(true)
+            setTimeout(() => {
+                setEditError(false);
+              }, "1500");
+        }
     }
 
     function updateComment(e) {
@@ -37,19 +43,26 @@ function RatingCard({rating, song, setRatings, listener}) {
     function updateRating(e) {
         e.preventDefault();
         e.target.children[0].value = ''
-
-        fetch(`/songs/${song.id}/ratings/${rating.id}`, {
-            method: 'PATCH',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({comment: commentChange}), 
-        }).then(resp => resp.json())
-        .then((newRating) => {
-            console.log(newRating)
-            const newRatings = song.ratings.map(rating => rating.id == newRating.id? newRating : rating) 
-            setRatings(newRatings)
-        })
+        if (listener.id == rating.id) {
+            fetch(`/songs/${song.id}/ratings/${rating.id}`, {
+                method: 'PATCH',
+                headers: {
+                    'Content-Type': 'application/json',
+                },
+                body: JSON.stringify({comment: commentChange}), 
+            }).then(resp => resp.json())
+            .then((newRating) => {
+                console.log(newRating)
+                const newRatings = song.ratings.map(rating => rating.id == newRating.id? newRating : rating) 
+                setRatings(newRatings)
+            })
+        }
+        else {
+            setEditError(true)
+            setTimeout(() => {
+                setEditError(false);
+              }, "1500");
+        }
     }
     
     return(
@@ -62,6 +75,7 @@ function RatingCard({rating, song, setRatings, listener}) {
                 <input type="submit" value="Change Comment"></input>
             </form>
             <button id="delete" onClick={handleDelete}>X</button>
+            <h5 style={{color: "red"}}>{editError ? "Listeners can only edit their own ratings" :null}</h5>
         </div>
     )
 }
